Downgrade vlc to 2.1.x from official repositories.
There is currently no APT repository out there that provides an easy way to install a newer version of VLC as library (libvlc) which is needed for WebChimera.js. Since we already have youtube-dl integrated, we no longer need to rely on VLC's functionality to resolve YouTube links, so in this case we can safely downgrade back to 2.1.x to get the library back.