Do not require an isolated GUI anymore.
This commit allows the TS3Bot to boot in an environment where
there is no way to boot an isolated graphical environment. In that case
TS3Bot will try to use the already existing display and the desktop
manager that runs on it already.

This adds some undefined error cases which still need to be found and
fixed, so be warned: This commit is very, VERY experimental!